move apply_overlay to image processor
